# app/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User Management System",
    "description": "Create and Manage Users, Roles and Permissions",
    "contact": {
      "name": "Bibek Aryal"
    },
    "license": {
      "name": "Personal Use Only"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/ums",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/app_users/user/{id}": {
      "get": {
        "tags": [
          "Users Management"
        ],
        "summary": "Retrieve app user by user ID",
        "description": "Fetches details of a specific app user by user ID.",
        "operationId": "readAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the specified app user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User Not Found by specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users Management"
        ],
        "summary": "Update app user details",
        "description": "Updates details of a specific app user by ID.",
        "operationId": "updateAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Updated details of the app user",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated details of the app user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User to update Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users Management"
        ],
        "summary": "Soft delete an app user",
        "description": "Marks a permission as deleted without permanently removing it, identified by its unique ID",
        "operationId": "softDeleteAppUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the app user to soft delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Not Authorized or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User to soft delete Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/app_users_roles/user_role": {
      "post": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Assign Role to a User",
        "description": "Adds a Role to a User",
        "operationId": "createAppUserRole",
        "requestBody": {
          "description": "Details of user and role to be linked",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppUserRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role assigned successfully to user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Required Element Missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseStatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User or Role to Link Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/apps_app_user/app/{appId}/user/{appUserEmail}": {
      "get": {
        "tags": [
          "Apps Users Management"
        ],
        "summary": "Get a linked app user by app ID and user email",
        "description": "Retrieves a specific app user its app ID and user email",
        "operationId": "readAppsAppUsersByAppIdAndUserEmail",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "description": "ID of the app in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "appUserEmail",
            "in": "path",
            "description": "Email of the User in linked app user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - App User Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsAppUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
    "/api/v1/app_users_roles": {
      "get": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Get all users and roles linked",
        "description": "Retrieves all linked users and roles available in the system",
        "operationId": "readAppUserRoles",
        "responses": {
          "200": {
            "description": "Users Roles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_users_roles/users/{userIds}": {
      "get": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Get linked users and roles by a list of user IDs",
        "description": "Retrieves all users and roles associated with the specified user IDs",
        "operationId": "readAppUserRolesByUserIds",
        "parameters": [
          {
            "name": "userIds",
            "in": "path",
            "description": "IDs of the users for which users roles are retrieved",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Apps Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_users_roles/user_role/{userId}/{roleId}": {
      "get": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Get a linked user role by user ID and role ID",
        "description": "Retrieves a specific user role by its user ID and role ID",
        "operationId": "readAppUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user in linked user role",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "description": "ID of the role in linked user role",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User Role Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Unassign Role from User",
        "description": "Removes a Role from User",
        "operationId": "deleteAppUserRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the User in linked user role",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roleId",
            "in": "path",
            "description": "ID of the Role in linked user role",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role unassigned successfully from user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - User Role Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
    "/api/v1/app_users_roles/user/{userId}": {
      "get": {
        "tags": [
          "Users Roles Management"
        ],
        "summary": "Get linked users and roles by user ID",
        "description": "Retrieves all users and roles associated with the specified user ID",
        "operationId": "readAppUserRolesByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user for which users roles are retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users Roles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User Not Authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User Forbidden or Not Validated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Other Errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppUserRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppsDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "ResponseCrudInfo": {
        "type": "object",
        "properties": {
          "insertedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "updatedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "deletedRowsCount": {
            "type": "integer",
            "format": "int32"
          },
          "restoredRowsCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponsePageInfo": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResponseStatusInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "errMsg": {
            "type": "string"
          }
        }
      },
      "AppUserAddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "addressType": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "AppUserRequest": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "status"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "guestUser": {
            "type": "boolean"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserAddressDto"
            }
          }
        }
      },
...
      "AppUserDto": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "status"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "guestUser": {
            "type": "boolean"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserAddressDto"
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "deletedDate": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppRoleDto"
            }
          },
          "validated": {
            "type": "boolean"
          }
        }
      },
      "AppUserResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserDto"
            }
          }
        }
      },
...
      "AppsAppUserDto": {
        "type": "object",
        "properties": {
          "app": {
            "$ref": "#/components/schemas/AppsDto"
          },
          "user": {
            "$ref": "#/components/schemas/AppUserDto"
          }
        }
      },
      "AppsAppUserResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "appsUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppsAppUserDto"
            }
          }
        }
      },
      "AppUserRoleRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AppUserRoleDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AppUserDto"
          },
          "role": {
            "$ref": "#/components/schemas/AppRoleDto"
          }
        }
      },
      "AppUserRoleResponse": {
        "type": "object",
        "properties": {
          "responseCrudInfo": {
            "$ref": "#/components/schemas/ResponseCrudInfo"
          },
          "responsePageInfo": {
            "$ref": "#/components/schemas/ResponsePageInfo"
          },
          "responseStatusInfo": {
            "$ref": "#/components/schemas/ResponseStatusInfo"
          },
          "usersRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppUserRoleDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Token": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "Basic": {
        "type": "http",
        "scheme": "Basic"
      }
    }
  }
}